Support Kotlin functional references in event subscribing
